hh-44,46: fix comments
