+++ app/src/main/java/ru/practicum/android/diploma/filter/domain/SharedPrefInteractorImpl.kt
@@ -1,31 +1,21 @@
 package ru.practicum.android.diploma.filter.domain
 
-import android.content.SharedPreferences
-import com.google.gson.Gson
+import ru.practicum.android.diploma.network.data.VacancyNetworkConvertor.convertToFilterIndustry
+import ru.practicum.android.diploma.network.data.VacancyNetworkConvertor.convertToFilterIndustryDto
 import ru.practicum.android.diploma.network.domain.models.FilterIndustry
 
 class SharedPrefInteractorImpl(
-    private val sharedPreferences: SharedPreferences,
-    private val gson: Gson
+    private val sharedPreferences: SharedPreferencesRepository
 ) : SharedPrefInteractor {
     override fun getChosenIndustry(): FilterIndustry {
-        val res = gson.fromJson(
-            sharedPreferences.getString(INDUSTRY_TAG, gson.toJson(DEFAULT_INDUSTRY_JSON)),
-            FilterIndustry::class.java
-        )
-        return res
+        return sharedPreferences.getChosenIndustry().convertToFilterIndustry()
     }
 
     override fun setIndustry(industry: FilterIndustry?) {
-        sharedPreferences.edit().putString(INDUSTRY_TAG, gson.toJson(industry)).apply()
+        sharedPreferences.setIndustry(industry.convertToFilterIndustryDto())
     }
 
     override fun resetIndustry() {
-        sharedPreferences.edit().putString(INDUSTRY_TAG, gson.toJson(DEFAULT_INDUSTRY_JSON)).apply()
-    }
-
-    companion object {
-        const val INDUSTRY_TAG = "INDUSTRY_TAG"
-        private val DEFAULT_INDUSTRY_JSON = FilterIndustry(-1, "")
+        sharedPreferences.resetIndustry()
     }
 }
