+++ app/src/main/java/ru/practicum/android/diploma/filter/domain/SharedPreferencesRepository.kt
@@ -1,0 +1,12 @@
+package ru.practicum.android.diploma.filter.domain
+
+import ru.practicum.android.diploma.network.data.dto.response.FilterIndustryDto
+
+interface SharedPreferencesRepository {
+
+    fun getChosenIndustry(): FilterIndustryDto
+
+    fun setIndustry(industry: FilterIndustryDto?)
+
+    fun resetIndustry()
+}
